requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy[asyncio]
aiomysql
python-jose
passlib[bcrypt]
pydantic
python-dotenv
